[metainfo] add possible better way to build our structs
Instead of all in metainfo, let each struct worry about how to build
itself from a hashmap.  This means other code worries about the IOError
and DecodeError of TorrentParserError, so only options are necessary,
although a more complex error type could be used to point to what/where
things are missing.  Since FormatError covers all of that might as well
just use an Option for now.

MultiFile is unimplemented, however, because there probably is a better
way to represent it as multiple SingleFiles rather than a different
struct

No tests, but it does work with the ubuntu iso torrent