[tracker-arch] Experiment with actually using the peer task
Adds an experiment feature, to run: `cargo run --features experiment`

This handshakes a peer, then sends a choke message if successful.
Gracefully closes after 5 seconds OR if peer disconnects, due to select
instead of join.

For local testing, I run the tokio echo example (so that it will
reciprocate the handshake and the simple message)

Next experiment: invert some control by returing the future instead
of spawning, and creating the peer output channel in main so that it
could be reused by multiple peers.  And an event loop to react to
messages