[channels] WIP Begin outlining real structure
Mainly with the comments in the --features real main and the listener,
which necessitated changes in Peer

Checks out but that's it

Peer needs an overhaul, a wrapper struct with the tx and the id does not
feel right, lifecycle start should return those two but then they
should be split up by main and put in the peer_map.  Also will require
an extension trait for the tx that'll give us nice methods like Peer
does currently.  Lastly the messages coming in need to be paired with a
PeerId, and the arguments to handle_stream need to be fixed up